Cutting Progress, World Canvas lesson completed
Created more events for the animation. Still finding it a bit hard to understand how events and subscribers work.

Where should the event go?

Code is getting more complex.

Created a counter for the number of interactions, an explored that part. Will be useful for building something like business dude.