Halfway through the Stove Counter, State machine lecture
So far added a new kind of scriptable object (burning) despite I don't know why we're doing it like that as when one ingredient burns it always burns to the other, so potentially we could have a base class and then cooking could be a child scriptable object with more things, so 5 parameters, in, cooked, burned, time to cook and time to burn.

Why is it not a good idea?