Completed lecture Plate Pick up objects
added the logic both ways, so if the player is carrying a plate we can add objects into the plate, and if there-s a plate in a counter, and the player carries objects, we can add those.

I do get the feeling that there's some code duplication and that it is a bit difficult to follow.

I have the feeling I could be using the this. to be clearer at some points, and that states could be a bit clearer as well.