Plate Complete Visual lecture completed
I really liked how he hacked the way of showing or not ingredients starting from a burger that has all of them and then only activating  the visuals.

The struct solution still not my favourite as it has a lot of setup done in the editor which I don't love.

What works better? A struct or a class? Maybe a class would use less memory, since now every time we're spawning a plate it comes with all of the struct used in memory.