Plate Complete visual
completed the plate visual. Pretty cool

Interesting to be thinking about how adding serializable Image from unity is also a gameObject and I'm adding the reference to a Image object not to a specific image itself so it can be setUp easily.

It looks like this also could be sorted, but it's more for a polish state.

Really powerful how to be using events here.